I created the basic models with SwiftData's @Model macro and made them conform to the Codable protocol.
I created the basic data models and added the SwiftData's macro @Model to make the data, when downloaded, persist on the device as if it were cache storage. I also added Codable protocol compliance to the models, so that we can make requests to our API from the data provided by the model itself.